+++ echoframework/src/main/kotlin/org/echo/mobile/framework/facade/SubscriptionFacade.kt
@@ -54,6 +54,8 @@ interface SubscriptionFacade {
      */
     fun subscribeOnContractLogs(
         contractId: String,
+        fromBlock: String,
+        limit: String,
         listener: UpdateListener<List<Log>>,
         callback: Callback<Boolean>
     )