+++ echoframework/src/main/kotlin/org/echo/mobile/framework/model/BitassetOptions.kt
@@ -5,7 +5,6 @@ import com.google.gson.JsonElement
 import com.google.gson.JsonObject
 import com.google.gson.annotations.SerializedName
 import org.echo.mobile.framework.ECHO_ASSET_ID
-import org.echo.mobile.framework.support.Uint16
 import org.echo.mobile.framework.support.Uint32
 import org.echo.mobile.framework.support.Uint8
 
@@ -18,9 +17,6 @@ import org.echo.mobile.framework.support.Uint8
 class BitassetOptions @JvmOverloads constructor(
     @SerializedName(FEED_LIFETIME_KEY) var feedLifetimeSec: Int = -1,
     @SerializedName(MINIMUM_FEED_KEY) var minimumFeeds: Int = 0,
-    @SerializedName(FORCE_SETTLEMENT_DELAY_KEY) var forceSettlementDelaySec: Int = -1,
-    @SerializedName(FORCE_SETTLEMENT_OFFSET_KEY) var forceSettlementOffsetPercent: Short = -1,
-    @SerializedName(MAXIMUM_FORCE_SETTLEMENT_VOLUME_KEY) var maximumForceSettlementVolume: Short = -1,
     @SerializedName(SHORT_BACKING_ASSET_KEY) var shortBackingAsset: String = ECHO_ASSET_ID
 ) : GrapheneSerializable {
 
@@ -30,15 +26,10 @@ class BitassetOptions @JvmOverloads constructor(
     override fun toBytes(): ByteArray {
         val feedLifetimeSecBytes = Uint32.serialize(feedLifetimeSec)
         val minimumFeedsBytes = Uint8.serialize(minimumFeeds)
-        val forceSettlementDelaySecBytes = Uint32.serialize(forceSettlementDelaySec)
-        val forceSettlementOffsetPercentBytes = Uint16.serialize(forceSettlementOffsetPercent)
-        val maximumForceSettlementVolumeBytes = Uint16.serialize(maximumForceSettlementVolume)
         val shortBackingAssetBytes = Uint8.serialize(Asset(shortBackingAsset).instance)
         val extensionsBytes = extensions.toBytes()
         return Bytes.concat(
-            feedLifetimeSecBytes, minimumFeedsBytes, forceSettlementDelaySecBytes,
-            forceSettlementOffsetPercentBytes, maximumForceSettlementVolumeBytes,
-            shortBackingAssetBytes, extensionsBytes
+            feedLifetimeSecBytes, minimumFeedsBytes, shortBackingAssetBytes, extensionsBytes
         )
     }
 
@@ -48,10 +39,6 @@ class BitassetOptions @JvmOverloads constructor(
         JsonObject().apply {
             addProperty(FEED_LIFETIME_KEY, feedLifetimeSec)
             addProperty(MINIMUM_FEED_KEY, minimumFeeds)
-            addProperty(FORCE_SETTLEMENT_DELAY_KEY, forceSettlementDelaySec)
-            addProperty(FORCE_SETTLEMENT_OFFSET_KEY, forceSettlementOffsetPercent)
-            addProperty(MAXIMUM_FORCE_SETTLEMENT_VOLUME_KEY, maximumForceSettlementVolume)
-            addProperty(MAXIMUM_FORCE_SETTLEMENT_VOLUME_KEY, maximumForceSettlementVolume)
             addProperty(SHORT_BACKING_ASSET_KEY, shortBackingAsset)
             add(EXTENSIONS_KEY, extensions.toJsonObject())
         }
@@ -59,9 +46,6 @@ class BitassetOptions @JvmOverloads constructor(
     companion object {
         private const val FEED_LIFETIME_KEY = "feed_lifetime_sec"
         private const val MINIMUM_FEED_KEY = "minimum_feeds"
-        private const val FORCE_SETTLEMENT_DELAY_KEY = "force_settlement_delay_sec"
-        private const val FORCE_SETTLEMENT_OFFSET_KEY = "force_settlement_offset_percent"
-        private const val MAXIMUM_FORCE_SETTLEMENT_VOLUME_KEY = "maximum_force_settlement_volume"
         private const val SHORT_BACKING_ASSET_KEY = "short_backing_asset"
         private const val EXTENSIONS_KEY = "extensions"
     }